requirements.txt
# Core dependencies
openai>=0.28.0
python-dotenv>=1.0.0
requests>=2.31.0
click>=8.1.7
rich>=13.6.0

# Web application
flask>=2.3.3
Werkzeug>=2.3.7
itsdangerous>=2.1.2
Jinja2>=3.1.2

# Utilities
pyyaml>=6.0.1
markupsafe>=2.1.3
